moved solution executing and planning from Grid class to PlannerController class
